add dark mode toggle for site
