+++ src/routes/+page.svelte
@@ -1,37 +1,24 @@
 <script>
-	import { settings } from '$lib/stores/settings.js'
-
-	function toggleDarkMode() {
-		settings.toggleDarkMode()
-	}
+	import { ROUTES } from '$lib/config.js'
 </script>
 
-<div class="w-full p-5">
-	<button class="float-right text-3xl dark:text-slate-200" on:click={toggleDarkMode}>
-		{#if $settings.darkMode}
-			<i class="fa-regular fa-lightbulb"></i>
-		{:else}
-			<i class="fa-regular fa-moon"></i>
-		{/if}
-	</button>
-</div>
-
-<div class="relative isolate px-6 pt-14 lg:px-8 -z-10">
-	<div class="absolute inset-x-0 -top-40 transform-gpu overflow-hidden blur-3xl sm:-top-80" aria-hidden="true">
+<div class="relative isolate px-6 pt-14 lg:px-8">
+	<div class="absolute inset-x-0 -top-40 transform-gpu overflow-hidden blur-3xl sm:-top-80 -z-10" aria-hidden="true">
 		<div
 			class="relative left-[calc(50%-11rem)] aspect-[1155/678] w-[36.125rem] -translate-x-1/2 rotate-[30deg] bg-gradient-to-tr from-[#ff80b5] to-[#9089fc] opacity-30 sm:left-[calc(50%-30rem)] sm:w-[72.1875rem]"
 			style="clip-path: polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)">
-		</div>
-	</div>
-	<div class="mx-auto max-w-2xl py-32 sm:py-48 lg:py-56">
-		<div class="text-center">
-			<h1 class="text-4xl font-bold tracking-tight text-gray-900 dark:text-slate-200 sm:text-6xl">JuicyJah</h1>
 		</div>
 	</div>
 	<div class="absolute inset-x-0 top-[calc(100%-13rem)] -z-10 transform-gpu overflow-hidden blur-3xl sm:top-[calc(100%-30rem)]" aria-hidden="true">
 		<div
 			class="relative left-[calc(50%+3rem)] aspect-[1155/678] w-[36.125rem] -translate-x-1/2 bg-gradient-to-tr from-[#ff80b5] to-[#9089fc] opacity-30 sm:left-[calc(50%+36rem)] sm:w-[72.1875rem]"
 			style="clip-path: polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)">
+		</div>
+	</div>
+	<div class="mx-auto max-w-2xl py-32 sm:py-48 lg:py-56 z-10">
+		<div class="flex flex-col justify-center items-center">
+			<h1 class="text-4xl font-bold tracking-tight text-center text-gray-900 dark:text-slate-200 sm:text-6xl p-2">JuicyJah</h1>
+			<a type="button" href={ROUTES.projects} class="button primary text-lg px-3 py-1 mt-4">See Projects <i class="ml-2 fa-solid fa-arrow-right"></i></a>
 		</div>
 	</div>
 </div>
